+++ next-docs/public/examples/table/ExpandedRow.tsx
@@ -91,22 +91,22 @@ const Example = () => {
   ];
 
   const makeData = (length: number) => {
-    return Array.from('_'.repeat(length)).map(() => {
+    return Array.from('_'.repeat(length)).map((_, index) => {
       return {
         firstName: 'Test',
-        age: <span>{Math.floor(Math.random() * 30)}</span>,
-        visits: <span>{Math.floor(Math.random() * 100)}</span>,
-        progress: <span>{Math.floor(Math.random() * 100)}</span>,
-        status: Math.floor(Math.random() * 100),
-        activity: Math.floor(Math.random() * 100),
+        age: <span>{Math.floor(index * 30)}</span>,
+        visits: <span>{Math.floor(index * 100)}</span>,
+        progress: <span>{Math.floor(index * 100)}</span>,
+        status: Math.floor(index * 100),
+        activity: Math.floor(index * 100),
         subRows: [
           {
             firstName: 'Sub Test',
-            age: <span>{Math.floor(Math.random() * 30)}</span>,
-            visits: <span>{Math.floor(Math.random() * 100)}</span>,
-            progress: <span>{Math.floor(Math.random() * 100)}</span>,
-            status: Math.floor(Math.random() * 100),
-            activity: Math.floor(Math.random() * 100),
+            age: <span>{Math.floor(index * 30)}</span>,
+            visits: <span>{Math.floor(index * 100)}</span>,
+            progress: <span>{Math.floor(index * 100)}</span>,
+            status: Math.floor(index * 100),
+            activity: Math.floor(index * 100),
           },
         ],
       };