+++ next-docs/tailwind.config.js
@@ -1,5 +1,6 @@
 module.exports = {
   content: [
+    './public/**/*.{js,ts,jsx,tsx}',
     './pages/**/*.{js,ts,jsx,tsx}',
     './components/**/*.{js,ts,jsx,tsx}',
   ],