+++ next-docs/CHANGELOG.md
@@ -1,5 +1,88 @@
 # Change Log
 
+## 7.20.0
+
+### Minor Changes
+
+- add moon-select package
+
+### Patch Changes
+
+- Updated dependencies [undefined]
+  - @heathmont/moon-select@7.20.0
+  - @heathmont/moon-accordion@7.20.0
+  - @heathmont/moon-assets@7.20.0
+  - @heathmont/moon-charts@7.20.0
+  - @heathmont/moon-components@7.20.0
+  - @heathmont/moon-core@7.20.0
+  - @heathmont/moon-draggabletable@7.20.0
+  - @heathmont/moon-icons@7.20.0
+  - @heathmont/moon-pagination@7.20.0
+  - @heathmont/moon-popover@7.20.0
+  - @heathmont/moon-table@7.20.0
+  - @heathmont/moon-themes@7.20.0
+  - @heathmont/moon-utils@7.20.0
+
+## 7.19.1
+
+### Patch Changes
+
+- Bump version
+- Updated dependencies [undefined]
+  - @heathmont/moon-accordion@7.19.1
+  - @heathmont/moon-assets@7.19.1
+  - @heathmont/moon-charts@7.19.1
+  - @heathmont/moon-components@7.19.1
+  - @heathmont/moon-core@7.19.1
+  - @heathmont/moon-draggabletable@7.19.1
+  - @heathmont/moon-icons@7.19.1
+  - @heathmont/moon-pagination@7.19.1
+  - @heathmont/moon-popover@7.19.1
+  - @heathmont/moon-table@7.19.1
+  - @heathmont/moon-themes@7.19.1
+  - @heathmont/moon-utils@7.19.1
+
+## 7.19.0
+
+### Minor Changes
+
+- Test coverage
+
+### Patch Changes
+
+- Updated dependencies [undefined]
+  - @heathmont/moon-accordion@7.19.0
+  - @heathmont/moon-assets@7.19.0
+  - @heathmont/moon-charts@7.19.0
+  - @heathmont/moon-components@7.19.0
+  - @heathmont/moon-core@7.19.0
+  - @heathmont/moon-draggabletable@7.19.0
+  - @heathmont/moon-icons@7.19.0
+  - @heathmont/moon-pagination@7.19.0
+  - @heathmont/moon-popover@7.19.0
+  - @heathmont/moon-table@7.19.0
+  - @heathmont/moon-themes@7.19.0
+  - @heathmont/moon-utils@7.19.0
+
+## 7.18.1
+
+### Patch Changes
+
+- Text Input sizes
+- Updated dependencies [undefined]
+  - @heathmont/moon-core@7.18.1
+  - @heathmont/moon-accordion@7.18.1
+  - @heathmont/moon-assets@7.18.1
+  - @heathmont/moon-charts@7.18.1
+  - @heathmont/moon-components@7.18.1
+  - @heathmont/moon-draggabletable@7.18.1
+  - @heathmont/moon-icons@7.18.1
+  - @heathmont/moon-pagination@7.18.1
+  - @heathmont/moon-popover@7.18.1
+  - @heathmont/moon-table@7.18.1
+  - @heathmont/moon-themes@7.18.1
+  - @heathmont/moon-utils@7.18.1
+
 ## 7.18.0
 
 ### Minor Changes