+++ packages/components/CHANGELOG.md
@@ -1,5 +1,51 @@
 # Change Log
 
+## 7.20.0
+
+### Minor Changes
+
+- add moon-select package
+
+### Patch Changes
+
+- Updated dependencies [undefined]
+  - @heathmont/moon-assets@7.20.0
+  - @heathmont/moon-themes@7.20.0
+  - @heathmont/moon-utils@7.20.0
+
+## 7.19.1
+
+### Patch Changes
+
+- Bump version
+- Updated dependencies [undefined]
+  - @heathmont/moon-assets@7.19.1
+  - @heathmont/moon-themes@7.19.1
+  - @heathmont/moon-utils@7.19.1
+
+## 7.19.0
+
+### Minor Changes
+
+- Test coverage
+
+### Patch Changes
+
+- Updated dependencies [undefined]
+  - @heathmont/moon-assets@7.19.0
+  - @heathmont/moon-themes@7.19.0
+  - @heathmont/moon-utils@7.19.0
+
+## 7.18.1
+
+### Patch Changes
+
+- Text Input sizes
+- Updated dependencies [undefined]
+  - @heathmont/moon-assets@7.18.1
+  - @heathmont/moon-themes@7.18.1
+  - @heathmont/moon-utils@7.18.1
+
 ## 7.18.0
 
 ### Minor Changes